+++ src/bibleref/__init__.py
@@ -21,20 +21,23 @@ symmetric difference).
 # Examples
 
 ```python
+>>> import bibleref
 >>> from bibleref import *
->>> # Parse a string of Bible ranges...
+>>>                             # Parse from string
 >>> range_list = BibleRangeList("Mark 2-3:6; 4; 6:1-6, 30-44, 56; Luke 2")
->>> print(range_list)               # Convert back to string
+>>> print(range_list)           # Convert back to a string
 Mark 2-3:6; 4; 6:1-6, 30-44, 56; Luke 2
 >>> len(range_list)
 6
->>> range_list[0]                   # Indiv ranges from a list
+>>> range_list[0]               # Individual ranges from a list
 BibleRange(Mark 2-3:6)
->>> range_list[5].start             # Start and...
+>>> range_list[1]
+BibleRange(Mark 4)
+>>> range_list[5].start         # Start and...
 BibleVerse(Luke 2:1)
->>> range_list[5].end               # ...end verses of a range.
+>>> range_list[5].end           # ...end verses of a range.
 BibleVerse(Luke 2:52)
->>> range_list[5].end.book          # Verse attributes
+>>> range_list[5].end.book      # Verse attributes
 <BibleBook.Luke: 'Luke'>
 >>> range_list[5].end.chap_num
 2
@@ -42,36 +45,50 @@ BibleVerse(Luke 2:52)
 52
 >>> len(range_list.groups)
 4
->>> range_list.groups[2]            # Range groups from a list
+>>> range_list.groups[2]        # Indivdual range groups from a list
 GroupView([BibleRange(Mark 6:1-6), BibleRange(Mark 6:30-44), BibleRange(Mark 6:56)])
->>> range_list.groups[2][0]         # Indiv ranges within the group
+>>> range_list.groups[2][0]     # Individual ranges within the group
 BibleRange(Mark 6:1-6)
 >>> range_list.groups[2][1]        
 BibleRange(Mark 6:30-44)
 >>> range_list.groups[2][2]
 BibleRange(Mark 6:56)
->>> BibleVerse('Mark 2:23') + 10    # Verse addition / subtraction
+>>> BibleVerse('Mark 2:23') + 10 # Verse addition / subtraction
 BibleVerse(Mark 3:5)
->>> BibleRange('1 John').split(by_chap=True, num_verses=10) # Range splits
-BibleRangeList("1 John 1, 2:1-10, 11-20, 21-29, 3:1-10, 11-20, 21-24, 4:1-10, 11-20, 21, 5:1-10, 11-20, 21")
->>> for verse in BibleRange('Mark 6:1-3'): 
-...     print(verse)                # Range iteration
+>>> BibleVerse('Mark 3:5') - BibleVerse('Mark 2:23')
+10
+>>> BibleRange('1 John').split(by_chap=True, num_verses=15) # Range splits
+BibleRangeList("1 John 1, 2:1-15, 16-29, 3:1-15, 16-24, 4:1-15, 16-21, 5:1-15, 16-21")
+>>> for verse in BibleRange('Mark 6:1-3'): # Range iteration
+...     print(verse)
 ... 
 Mark 6:1
 Mark 6:2
 Mark 6:3
->>> BibleRange('Matt 2:3-John 4:5', flags=BibleFlag.MULTIBOOK) # Multibook ranges
-BibleRange(Matthew 2:3-John 4:5)
+>>> bibleref.flags = BibleFlag.MULTIBOOK   # Enable multi-book ranges
+>>> BibleRange('Matt 10-John 10')
+BibleRange(Matthew 10-John 10)
 >>> list_1 = BibleRangeList("Matt 2-4; Mark 6-8; Luke 10-12; John 14-16")
->>> list_2 = BibleRangeList("John 1-3; Luke 9-11, 13; Matt 3-5; Mark 12")
->>> list_1 | list_2                 # Union of range lists
-BibleRangeList("Matthew 2-5, Mark 6-8, 12, Luke 9-13, John 1-3, 14-16")
->>> list_1 & list_2                 # Intersection range lists
-BibleRangeList("Matthew 3-4, Luke 10-11")
->>> list_1 - list_2                 # Difference of range lists
-BibleRangeList("Matthew 2, Mark 6-8, Luke 12, John 14-16")
->>> list_1 ^ list_2                 # Symmetric difference of range lists
-BibleRangeList("Matthew 2, 5, Mark 6-8, 12, Luke 9, 12-13, John 1-3, 14-16")
+>>> list_2 = BibleRangeList("John 1-3; Luke 9; Matt 3-5; Mark 12")
+>>> list_1 | list_2                     # Union
+BibleRangeList("Matthew 2-5; Mark 6-8; 12; Luke 9-12; John 1-3; 14-16")
+>>> list_1 & list_2                     # Intersection
+BibleRangeList("Matthew 3-4")
+>>> list_1 - list_2                     # Difference
+BibleRangeList("Matthew 2; Mark 6-8; Luke 10-12; John 14-16")
+>>> list_1 ^ list_2                     # Symmetric difference
+BibleRangeList("Matthew 2; 5; Mark 6-8; 12; Luke 9-12; John 1-3; 14-16")
+>>> range_list = BibleRangeList("Mark 3:2-4:5; 1 John 1:5-3 John 8;")
+>>> range_list.verse_count()            # Count of verses
+161
+>>> range_list.chap_count()             # Count of chapters (incl partial)
+9
+>>> range_list.chap_count(whole=True)   # Count of chapters (whole only)
+5
+>>> range_list.book_count()             # Count of books (incl partial)
+4
+>>> range_list.book_count(whole=True)   # Count of books (whole only)
+1
 ```
 
 # Attribution