+++ CHANGES.md
@@ -1,5 +1,6 @@
 Next version:
   - Add counts to BibleBook, BibleRange, BibleRangeList
-  
+  - Add BibleRangeList.regroup(), and call it from consolidate()
+  - Fix bug around setting bibleref.flags  
 v0.9.0:
   - Initial release